Fix template updating
When loading template A with a dependency of B, if B was modified then A would update. However, template C with a dependency of B would not update. Now when loading A and updating B, all templates than have B as a dependency are marked as modified and will be ready to update when loaded.