Add circle init feature
The config file `circles_init.json` is read on startup. It contains information needed to construct circles. For each circle, if it does not exist, then make a new one. If it does exist, modify its state in the database to match its state in the config file. The modification part isn't done yet, and neither part has been tested.